Banco de dados layout/requirements.txt
streamlit==1.47.0
pandas==2.1.3
numpy==1.24.3
matplotlib==3.7.2
plotly==5.17.0
seaborn==0.12.2
openpyxl==3.1.2
xlrd==2.0.1
ofxparse==0.21
python-dateutil==2.8.2
azure-storage-blob==12.19.0
azure-keyvault-secrets==4.7.0
azure-identity==1.15.0
azure-applicationinsights==0.11.10
sendgrid==6.10.0
psycopg2-binary==2.9.7
SQLAlchemy==2.0.23
streamlit-authenticator==0.2.3
bcrypt==4.1.2
jwt==1.3.1
requests==2.31.0
